% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_data.R
\name{acquire_data}
\alias{acquire_data}
\title{Get newest data from all packages and log any errors}
\usage{
acquire_data()
}
\value{
New data in the \code{data} dir.
}
\description{
Uses \code{/data/packages.csv} CSV to try the package's \code{get_info} function as well as its \code{refresh} function to get the most up to date data.
}
\examples{
\donttest{
acquire_data
}
}
