% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron_job_dependencies.R
\name{install_depend}
\alias{install_depend}
\title{Install Required covid19 Packages}
\usage{
install_depend(verbose = TRUE)
}
\arguments{
\item{verbose}{if TRUE, will show messages}
}
\description{
Install Required covid19 Packages
}
\details{
Install the required packages for the cron job
}
